olfati_django/requirements.txt
Django~=4.2.1
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.2.2
kavenegar~=1.1.2
Pillow~=9.5.0
psycopg2-binary~=2.9.6
celery~=5.4.0
django_celery_beat~=2.6.0
django_celery_results~=2.5.1
django_redis~=5.4.0
firebase_admin~=6.5.0